[department-of-veterans-affairs/#orchid-107] Reorganize test structure to differentiate between unit and cypress and match code structure
