+++ src/applications/ask-a-question/tests/cypress/ask-a-question.cypress.spec.js
@@ -3,8 +3,8 @@ import path from 'path';
 import testForm from 'platform/testing/e2e/cypress/support/form-tester';
 import { createTestConfig } from 'platform/testing/e2e/cypress/support/form-tester/utilities';
 
-import formConfig from '../form/config/form';
-import manifest from '../manifest.json';
+import formConfig from '../../form/config/form';
+import manifest from '../../manifest.json';
 
 // `appName`, `arrayPages`, and `rootUrl` don't need to be explicitly defined.
 const testConfig = createTestConfig(
